Modifiche allo script che effettua le predizioni

+++ predictor/less_sinful_baseline.py
@@ -89,7 +89,7 @@ class LessSinfulBaselinePredictor(object):
         if self.pcp_estimation is None:
             return None
 
-        predictions = np.zeros(shape=self.pcp_estimation.shape)
+        predictions = np.zeros(shape=self.pcp_estimation.shape, dtype=int)
         clients_count = predictions.shape[0]
         products_count = predictions.shape[1]
